moved `constructUri` to `utils.dart`
